+++ lib/dop_common/plan.rb
@@ -37,12 +37,12 @@ module DopCommon
 
     def max_in_flight
       @max_in_flight ||= max_in_flight_valid? ?
-        @hash[:plan][:max_in_flight] : DEFAULT_MAX_IN_FLIGHT
+        @hash[:max_in_flight] : DEFAULT_MAX_IN_FLIGHT
     end
 
     def ssh_root_pass
       @ssh_root_pass ||= ssh_root_pass_valid? ?
-        @hash[:plan][:ssh_root_pass] : nil
+        @hash[:ssh_root_pass] : nil
     end
 
     def infrastructures
@@ -81,18 +81,32 @@ module DopCommon
     end
 
     def max_in_flight_valid?
-      return false if @hash[:plan].nil? # plan hash is optional
-      return false if @hash[:plan][:max_in_flight].nil? # max_in_flight is optional
-      @hash[:plan][:max_in_flight].kind_of?(Fixnum) or
+      ### START DEPRICATED KEY PARSING plan => max_in_flight
+      if @hash[:max_in_flight].nil?
+        return false if @hash[:plan].nil? # plan hash is optional
+        return false if @hash[:plan][:max_in_flight].nil? # max_in_flight is optional
+        @hash[:max_in_flight] = @hash[:plan][:max_in_flight]
+        DopCommon.log.warn('The max_in_flight key under "plan" in depricated. Please set max_in_flight as a global key')
+      end
+      ### END DEPRICATED KEY PARSING
+      return false if @hash[:max_in_flight].nil? # max_in_flight is optional
+      @hash[:max_in_flight].kind_of?(Fixnum) or
         raise PlanParsingError, 'Plan: max_in_flight has to be a number'
-      @hash[:plan][:max_in_flight] > 0 or
-        raise PlanParsingError, 'Plan: max_in_flight has to be greater than one'
+      @hash[:max_in_flight] >= -1 or
+        raise PlanParsingError, 'Plan: max_in_flight has to be greater than -1'
     end
 
     def ssh_root_pass_valid?
-      return false if @hash[:plan].nil? # plan hash is optional
-      return false if @hash[:plan][:ssh_root_pass].nil? # ssh_root_pass is optional
-      @hash[:plan][:ssh_root_pass].kind_of?(String) or
+      ### START DEPRICATED KEY PARSING plan => max_root_pass
+      if @hash[:ssh_root_pass].nil? # ssh_root_pass is optional
+        return false if @hash[:plan].nil? # plan hash is optional
+        return false if @hash[:plan][:ssh_root_pass].nil? # ssh_root_pass is optional
+        @hash[:ssh_root_pass] = @hash[:plan][:ssh_root_pass]
+        DopCommon.log.warn('The ssh_root_pass key under "plan" in depricated. Please set ssh_root_pass as a global key')
+      end
+      ### END DEPRICATED KEY PARSING
+      return false if @hash[:ssh_root_pass].nil? # ssh_root_pass is optional
+      @hash[:ssh_root_pass].kind_of?(String) or
         raise PlanParsingError, 'Plan: ssh_root_pass has to be a string'
     end