+++ spec/dop_common/plan_spec.rb
@@ -31,6 +31,19 @@ describe DopCommon::Plan do
       expect(plan.max_in_flight).to be DopCommon::Plan::DEFAULT_MAX_IN_FLIGHT
     end
     it 'will return the correct value if max_in_flight is defined' do
+      plan = DopCommon::Plan.new({:max_in_flight => 10})
+      expect(plan.max_in_flight).to be 10
+    end
+    it 'will throw and exception if the value is not a Fixnum' do
+      plan = DopCommon::Plan.new({:max_in_flight => 'foo'})
+      expect{plan.max_in_flight}.to raise_error DopCommon::PlanParsingError
+    end
+    it 'will throw and exception if the value is < -1' do
+      plan = DopCommon::Plan.new({:max_in_flight => -2})
+      expect{plan.max_in_flight}.to raise_error DopCommon::PlanParsingError
+    end
+    ### START DEPRICATED KEY PARSING plan => max_in_flight
+    it 'will return the correct value if max_in_flight is defined' do
       plan = DopCommon::Plan.new({:plan => {:max_in_flight => 10}})
       expect(plan.max_in_flight).to be 10
     end
@@ -38,10 +51,11 @@ describe DopCommon::Plan do
       plan = DopCommon::Plan.new({:plan => {:max_in_flight => 'foo'}})
       expect{plan.max_in_flight}.to raise_error DopCommon::PlanParsingError
     end
-    it 'will throw and exception if the value is < 1' do
-      plan = DopCommon::Plan.new({:plan => {:max_in_flight => -1}})
+    it 'will throw and exception if the value is < -1' do
+      plan = DopCommon::Plan.new(:plan => {:max_in_flight => -2})
       expect{plan.max_in_flight}.to raise_error DopCommon::PlanParsingError
     end
+    ### END DEPRICATED KEY PARSING
   end
 
   describe '#ssh_root_pass' do
@@ -50,6 +64,15 @@ describe DopCommon::Plan do
       expect(plan.ssh_root_pass).to be nil
     end
     it 'will return the correct value if ssh_root_pass is defined' do
+      plan = DopCommon::Plan.new({:ssh_root_pass => 'mypass'})
+      expect(plan.ssh_root_pass).to eq 'mypass'
+    end
+    it 'will throw and exception if the value is not a String' do
+      plan = DopCommon::Plan.new({:ssh_root_pass => 2})
+      expect{plan.ssh_root_pass}.to raise_error DopCommon::PlanParsingError
+    end
+    ### START DEPRICATED KEY PARSING plan => ssh_root_pass
+    it 'will return the correct value if ssh_root_pass is defined' do
       plan = DopCommon::Plan.new({:plan => {:ssh_root_pass => 'mypass'}})
       expect(plan.ssh_root_pass).to eq 'mypass'
     end
@@ -57,6 +80,7 @@ describe DopCommon::Plan do
       plan = DopCommon::Plan.new({:plan => {:ssh_root_pass => 2}})
       expect{plan.ssh_root_pass}.to raise_error DopCommon::PlanParsingError
     end
+    ### END DEPRICATED KEY PARSING
   end
 
   describe '#infrastructures' do